Use Slate transforms for block merges, add matchMetadataFromDom to isContentReady
Block merge (backspace at start joining into previous block) now uses a
headless Slate editor with Editor.deleteBackward instead of manual JSON
manipulation. This produces valid normalized Slate (no adjacent empty text
nodes) and handles all merge cases: p-into-p, p-into-list, list-into-list.

Adjacent list merge also uses Slate (Transforms.mergeNodes).

isContentReady now uses matchMetadataFromDom mode: metadata values from
formData are only included in the comparison if they appear as DOM attribute
values. This correctly detects rendered changes (link URL updates) while
ignoring metadata the renderer does not render.

Also: handleDeleteKey supports fields where data-node-id equals data-edit-text
element, updateEmptyState skips focused fields, rename updateJsonNode.spec.ts
to readSlateValueFromDOM.spec.ts.
